feat: add instrumentation and counter for event (#238)
Resolves #205 

The approach is here to use [manual
configuration](https://nextjs.org/docs/app/guides/open-telemetry#manual-opentelemetry-configuration)
to allow more customisability.
A simple example of a counter for http is created.

This is tested with
https://github.com/vercel/opentelemetry-collector-dev-setup?tab=readme-ov-file
and the output is as seen on the screenshot.
<img width="1150" height="642" alt="Screenshot 2025-12-11 at 10 55
12 PM"
src="https://github.com/user-attachments/assets/7932845b-8b47-4d75-9662-92aa79a1cb87"
/>

Here, I have added
- a facade to interact with otel metrics
- a wrapper for API to run metrics

---------

Co-authored-by: Richard Boomsma <spaaacetoast@proton.me>